deepfake_detector/requirements.txt
fastapi==0.135.3
uvicorn[standard]==0.43.0
torch==2.11.0
transformers==5.5.0
librosa==0.11.0
numpy==2.4.4
pydantic==2.12.5
python-dotenv==1.2.2
python-multipart>=0.0.9
